requirements.txt
Django==3.2
psycopg2==2.7.7
dotenv==0.21.0
django-filter==22.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.6.0
djoser==2.1.0
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.4
Pillow==9.4.0
django-extra-fields==3.0.2